refactor(components): move forms into separate folder
