+++ src/components/forms/FormExample/FormExample.stories.ts
@@ -4,7 +4,7 @@ import { i18n } from '@/test/i18n';
 import ComponentDocs from '@/assets/design-system/docs/components/ComponentDocs.vue';
 import FormExample from './FormExample.vue';
 
-const story = storiesOf('Input & Actions|FormExample', module) as any;
+const story = storiesOf('Forms|FormExample', module) as any;
 
 story.add(
   'Default',
